Fetch default epic copy dynamically
Fetch the copy for the default epic from an API endpoint once when we
boot the app. At this point we're still missing the following, which
we'll follow up on in future PRs:

* Elegant handling/retrying if we fail to fetch the content

* We don't have full support for all possible placeholders yet (I think
the full list is %%CURRENCY_SYMBOL%%, %%COUNTRY_NAME%% and
%%ARTICLE_COUNT%%)

* We do support interpolating the currency symbol, but we just use a
hardcoded value. In future we'll want to bring across the logic from
frontend where we resolve this based on the geolocation.