Fetch default epic content lazily at request time
Previously we were fetching at boot time (asynchronously) and then
starting the app. This approach caused issues with the deployed lambda,
I think around the fact that the exports.handler call was happening
asynchronously. I looked into Amazon's docs but didn't quite find an
approach that worked. So instead we're now lazily fetching the epic
content when the request comes in (and caching the response), which
simplifies the server.

Note that I've also added a line to the server to kick off the async
default epic request immediately, which should reduce the waiting time
for the first request.