Regenerating JSON Schemas with Husky
This PR makes use of Git Hooks to regenerate the JSON schemas behind the scenes on pre-push and stop the push if there are any schema changes, so these can be reviewed and committed separately. If no changes detected, let the push go through but inform the user no changes were found. We use Husky to manage the hooks for us.