Disable JSON schema validation
At the moment this is configured via a const which is hardcoded to
false. I think going forward we could pull this value from configuration
in parameter store (or similar) so we can toggle it per environment.

We discussed not enabling in prod (it does after all have a small
performance penalty) but I can see it continuing to be useful in CODE
and DEV.

Another option in the short term might be to do the validation in PROD
but simply log the output (and don’t return a 400). This would allow us
to gather data on whether our understanding of the input params is
correct.