Add showSupportMessaging field
This term comes from:

https://github.com/guardian/members-data-api/blob/3a72dc00b9389968d91e5930686aaf34d8040c52/membership-attribute-service/app/models/Attributes.scala

and is referred to in Frontend as such, so we're reusing it.

Unfortunately, it is not sufficient:

https://github.com/guardian/frontend/blob/master/static/src/javascripts/projects/common/modules/commercial/user-features.js#L276

so we have to pass in some extra stuff too.

Ideally, we'd consolidate this on the platforms, but we don't want
to duplicate this logic across Frontend and DCR.

It's also debatable whether teams should need to change the Members
Data API to experiment with new definitions of this value in tests.