Primary CTA is optional
The primary CTA button is optional. If there is no cta key in the
variant data then we shouldn't show it. If it's there but any fields are
missing then we fall back on defaults.

Also includes a fix to only run tests in src. Before this the
EpicButtons tests were running in src (ts version) and dist (js version)
with the js version failing with a fairly cryptic error.  See
https://stackoverflow.com/questions/50145078/jest-typescript-tests-runs-twice-one-for-ts-files-and-one-for-js-files