Start using factories for test data
Using the fishery library. This doesn't quite cover all scenarios, but
the PR is getting quite big so fill follow up in future PRs to migrate
e.g. storybook over to using factories. The idea is that the factories
become the canonical place to go for creating valid test data.