Remote es modules spike (#138)
* Update yarn lock

* Serve Epic as ES Module

(This is a backwards compatible change.)

ES modules can be dynamically imported across domains. By serving
the epic as a module, React (or another framework) can be used to
build interactive epics. Alternative approaches - such as evaling
JS, or packaging as an NPM module - have considerable downsides.

We expect this approach to have broader application than just the
epic.

As part of this work, the API interface has been added to:

* a new GET /epic.js endpoint that return the module
* a new `dataOnly` parameter to return only metadata from POST /epic

Rollup has also been introduced as a bundler, chosen because it
can output ES modules out of the box, and also because it seems
simpler than Webpack to my untrained eyes.

Modules are generated at build time. For local development,
modules are generated on startup, but for watching use the new

    yarn modules-dev

command. (Run this in a separate tab.)

Further testing is required for the module approach on browser
support and performance. If successful, code related to the old
approach will be subsequently removed.

* Fix some lint warnings

* Attempt fix for Babel issue

Seems to be: https://github.com/babel/babel/issues/11216.

* Fix param type error

* Revert node version bump