Fix didRenderEpic value
Currently we're reporting that we rendered an epic 100% of the time,
since here response is always truthy. The response value here now has
the relevant epic data nested under a data property already, so it's
that we need to inspect to determine if an epic was "rendered" (that
term might need revisiting now we're switching to the data approach).