Add ticker support to the epic
This is a React version of the ticker behaviour on frontend. I've only
added support for the "unlimited" version (as opposed to "hardstop")
since that's what's needed currently.

Note that I've borrowed useHasBeenSeen from dcr. This is so that we can
trigger the animation when the ticker comes into the viewport. If this
approach proves successful we can extract useHasBeenSeen into an npm
package or something.

Ultimately we might want the ticker data to be fetched server side and
passed in as props, in which case we could remove the
ContributionsEpicTickerWithDataFetch component.

Note that currently this won't be used because there's a filter which
checks if a variant has a ticker and skips that test if it does. We
shouldn't remove that check until we're moved over to remote modules and
are able to fully support the interactivity of the ticker.