Perform basic check on render
Rather than add lots of code to check for unexpected placeholders,
let's just rely on a basic last-minute check in the rendering.

This guarantees we won't show placeholders to users.

The advantage is simplicity and less code to maintain.

The disadvantage is poor handling of unexpected placeholders - we
will still select variants with these and only fail at the render
stage when it is too late to choose another variant. In practice
though, placeholders don't get added very often and they can be
added here first when they do to avoid this.