Log user-agent string.
We've been investigating a spike in 500 responses from dotcom components.
It's tricky to tell where these requests are coming from so we've decided
to start logging the user agent string to see if that can provide any
clarity. The step change in 5xxs and the regularity of them is leading us
to think it might be coming from something internal.