Add temporary 'validation' to /header endpoint.
This change adds a `bodyContainsAllFields` middleware that we can
use on the /header endpoint. This provides a minimal amount of
validation that should be sufficient to address an on going alarm.
This alarm has been caused by POST requests to /header without a
valid request body. The proper solution is to add zod for validation
but that will require a refactor to the packages to avoid the
dotcom package pulling in zod as an unrequired dependency. This change
is just temporary and can be removed after implementing proper validation.