Add moment template banner component.
This change adds the `MomentTemplateBanner` component. The idea with
this component is to allow for quickly putting together moment banners
in the future. The ultimate goal would be to hook this up to the RRCP
to allow configuring moment banners without any dev involvment.

For now, a dev will have to create moment banners using this template.
This will involve importing the `getMomentTemplateBanner` function and
passing in some `BannerTemplateSettings` to set the right colours and
images.

In a follow up PR I will add the remiders component.