Add specific moment template banner article count opt out.
This change adds an article count opt out component specific to
the moment template banner (MomentTemplateBannerArticleCountOptOut).

This did involve a lot of copy and pasting from the existing opt out
(ArticleCountOptOutPopup) which is a shame but I think having some
duplication is probably the simplest solution for now. The issue is
that the existing component takes a `type` parameter which it then uses
to decide on the styles for the component. This doesn't quite fit with
moment template banner which we want to be able to fully customise just
through the `BannerTemplateSettings` that we provide. The settings already
specify everything that we need to style the opt out (background colour,
primary cta, secondary cta, and close button) so it was quite simple to
create a new component that accepts a `BannerTemplateSettings` instead of
an `ArticleCountOptOutType`.

I removed tracking from the component as when I checked the code it was
only being used for epics - no banners were passing in the tracking object.

It's worth noting that this new component is only used in the separate,
hardcoded article count subheading. If a test is configured in the RRCP
that uses the %%ARTICLE_COUNT%% tag in the body copy, the older version
is used (it will default to the yellow style of the main contribs banner).
This is actually an existing problem with all moment banners so felt out
of scope to fix now, especialy given the reason for having the subheading
is to remove the need to reference article count in the body text. The reason
this is quite tricky to solve is those %% tags are replaced in the shared
`bannerWrapper` so it will require a decent refactor to get right.