Allow highlighted text to be customised in the moment template banner.
This change adds `HighlightedTextSettings` to the `BannerTemplateSettings` type.
This allows specifying the text colour & highlight colour. I've implemented the
highlight using a box-shadow and `box-decoration-break` which is slightly different
to how we've done it previously. This allows for the begining and end of wrapped lines
to still have some padding. In older browsers that don't support `box-decoration-break`
it falls back to having the start and end of lines tight with the end of the highlight.
This looks less good visually, but I think it's a completely acceptable fallback.