Update bannerWrapper to handle falling back on mainContent when mobileContent is undefined.
This makes a small tweak to the `bannerWrapper` that handles using `mainContent` as
a fallback when `mobileContent` is undefined. By moving this step into the banner wrapper
it simplifies the logic in individual banners for rendering mobile vs desktop content.

I've updated the `ContributionsBanner` and `MomentTemplateBanner` components to handle this
behaviour. We could also simplify all the other banners, but as they are no longer in use,
it's probably not worth it (we should look into cleaning them up soon).