Display logo from tablet in choice cards banner
Based on the designs, this logo should be visible in the tablet
breakpoint and not in the mobile medium breakpoint: looks like we made a
mistake with the breakpoint when we made this in the first place.

Interestingly, this is a change from some other banners: the guardian
weekly and print, for example. Those both display the logo from
mobileMedium. Not having seen the designs, I don’t know whether that’s
correct, but it has been that way for a long time so I assume so – I’ll
get this out and then check.