Force simple-update-notifier to use semver@7.5.2
We're on the latest storybook. The @storybook/cli package relies on
simple-update-notifier@^1.0.0, the latest version which this can resolve
to is 1.1.0. This depends on semver@~7.0.0 which only resolves to
vulnerable versions. So, assume that semver practices semver and force
it to 7.5.2, which doesn't have the vulnerability!

The leading ** in the resolution was needed to get this working with
workspaces. I think it'll apply to all workspaces (but that's okay,
because only one of them has simple-update-notifier in the bundle).