Fix remaining semver vulnerabilities (#934)
I had to force simple-update-notifier to use semver@7.5.2 with a resolution.

The dependency chain is @storybook/cli > simple-update-notifier >
semver@~7.0.0.

I've upgraded to the latest storybook. The @storybook/cli package relies on
simple-update-notifier@^1.0.0, the latest version which this can resolve to is
1.1.0. This depends on semver@~7.0.0 which only resolves to vulnerable
versions. So, assume that semver practices semver and force it to 7.5.2, which
doesn't have the vulnerability!

The leading ** in the resolution was needed to get this working with
workspaces. I think it'll apply to all workspaces (but that's okay, because
only one of them has simple-update-notifier in the bundle).

Note that a warning is emitted when running yarn install: warning Resolution
field "semver@7.5.2" is incompatible with requested version "semver@~7.0.0"
This is expected.

After the upgrade I was seeing an error in the build relating to
@storybook/addon-knobs. I noticed that the majority of stories use controls
rather than knobs, so I migrated the remaining 3 stories which were still using
knobs to use controls. This then allows us to drop the @storybook/addon-knobs
dependency and the build error goes away.