The getBanner pattern doesn't make sense here
The DesignableBanner is a component we want to use directly, unlike the
momment template which defines a function you can call to _return_ a
component you can use directly.

There's some stuff going on with the wrapper I'm not 100% sure of. I've
made it work for the story but there may be more to do here.