Designable Banner (#950)
This PR adds a new banner component DesignableBanner which supports a limited
design being passed in as props. The design is configured in the RRCP.
Currently only the images are in scope for the design, but this will be
increased to cover all colours etc.

When a banner variant has a design (rather than a template) the design is
fetched from Dynamo (actually these are prefetched asynchronously, following
the existing pattern) and added to the props returned by SDC to be passed to
the component.