If loading tests fails, return a failed promise
Instead of swallowing the failure and returning an empty list log/record
a metric for the error and return a failed promise.

This function is used in ValueReloader instances which are designed to
use already cached data (previously loaded) in the case of a failure. By
swallowing errors and returning an empty list, SDC may lose previously
successfully loaded test data, and will therefore not return messages.

In the case that the initial load fails, the server won't initialise
correctly and so it won't be added to the ASG, causing Riff Raff to
mark the deploy as failed. I think being explicit here is also
desirable.