Add an alarm for banner design loading failures (#956)
Add a metric and alarm for banner design Dynamo loading failures.

For both tests and banner designs, if loading from Dynamo fails, return a
failed promise. Instead of swallowing the failure and returning an empty list
log/record a metric for the error and return a failed promise.

These functions are used in ValueReloader instances which are designed to use
already cached data (previously loaded) in the case of a failure. By swallowing
errors and returning an empty list, SDC may lose previously successfully loaded
test and banner design data, and will therefore potentially break messaging.

In the case that the initial load fails, the server won't initialise correctly
and so it won't be added to the ASG, causing Riff Raff to mark the deploy as
failed. I think being explicit here is also desirable.

Also fix an import in a test which TypeScript was complaining about. Not sure
why this wasn't caught previously.