Add Giving Tuesday and New Years Eve reminders logic (#993)
Start showing Remind Me on Giving Tuesday on the Epic CTA in the US. This will
continue until November 27th (the day before Giving Tuesday). Then on November
28th start showing "Remind me on New Years Eve" which will continue showing
until December 30th.

Each of these distinct reminder periods has it's own reminder_option:

giving-tuesday-2023
new-years-eve-2023

These are sent to the API and end up in Braze so can be used for targeting.

Note: this reminder logic will be used by the banner too, but the CTA is
hardcoded and therefore won't change.