Zod Upgrade (alternative attempt!) (#997)
Upgrade zod to address a security issue.

Strangely I had to update one of the imports in the shared package from `import
{ z } from 'zod'` to `import * as z from 'zod'`. This matches the other imports
of zod in shared. It's not clear to me why this was needed, but locally using
the named import I observed that z was undefined, which was causing errors.